+++ ALUMNOS/MDAB/CARLOS_OLIVER/E2E-CHUCK/docker-compose.yml
@@ -1,46 +1,20 @@
-services:
-    
-  nifi:
-    image: apache/nifi:1.27.0
-    container_name: nifi
-    restart: always
-    ports:
-      - 8443:8443
-    environment:
-      - SINGLE_USER_CREDENTIALS_USERNAME=admin
-      - SINGLE_USER_CREDENTIALS_PASSWORD=ctsBtRBKHRAx69EqUghvvgEvjnaLjFEB
-    volumes:
-      - ./data:/tmp/data
-    user: "0:0"
+version: '3.8'
 
-  postgres:
-      container_name: postgres_container_chuck
-      image: postgres:12.1
-      environment:
-        POSTGRES_USER: ${POSTGRES_USER:-postgres}
-        POSTGRES_PASSWORD: ${POSTGRES_PASSWORD:-Welcome01}
-        PGDATA: /data/postgres
-      volumes:
-        - postgres:/data/postgres
+services:
+  nifi:
+      image: apache/nifi:1.27.0
+      container_name: nifi
+      restart: always
       ports:
-        - "5432:5432"
+        - 8443:8443
+      environment:
+        - SINGLE_USER_CREDENTIALS_USERNAME=admin
+        - SINGLE_USER_CREDENTIALS_PASSWORD=ctsBtRBKHRAx69EqUghvvgEvjnaLjFEB
+      volumes:
+        - ./data:/tmp/data
+      user: "0:0"
       networks:
-        - postgres
-      restart: unless-stopped
-   
-  pgadmin:
-    container_name: pgadmin_container_chuck
-    image: dpage/pgadmin4:4.16
-    environment:
-      PGADMIN_DEFAULT_EMAIL: ${PGADMIN_DEFAULT_EMAIL:-pgadmin4@pgadmin.org}
-      PGADMIN_DEFAULT_PASSWORD: ${PGADMIN_DEFAULT_PASSWORD:-admin}
-    volumes:
-       - pgadmin:/root/.pgadmin
-    ports:
-      - "${PGADMIN_PORT:-5050}:80"
-    networks:
-      - postgres
-    restart: unless-stopped
+        - postgres  
 
   mongo:
     image: mongo
@@ -51,7 +25,9 @@ services:
       MONGO_INITDB_ROOT_USERNAME: root
       MONGO_INITDB_ROOT_PASSWORD: example
     volumes:
-      - data:/data/db
+      - mongo_data:/data/db
+    networks:
+      - postgres  
 
   mongo-express:
     image: mongo-express
@@ -62,12 +38,45 @@ services:
       ME_CONFIG_MONGODB_ADMINUSERNAME: root
       ME_CONFIG_MONGODB_ADMINPASSWORD: example
       ME_CONFIG_MONGODB_URL: mongodb://root:example@mongo:27017/
+    networks:
+      - postgres 
+
+  postgres:
+    container_name: postgres_container
+    image: postgres:12.1
+    environment:
+      POSTGRES_USER: ${POSTGRES_USER:-postgres}
+      POSTGRES_PASSWORD: ${POSTGRES_PASSWORD:-Welcome01}
+      PGDATA: /data/postgres
+    volumes:
+      - postgres:/data/postgres
+    ports:
+      - "5432:5432"
+    networks:
+      - postgres  
+    restart: unless-stopped
+
+  pgadmin:
+    container_name: pgadmin_container
+    image: dpage/pgadmin4:4.16
+    environment:
+      PGADMIN_DEFAULT_EMAIL: ${PGADMIN_DEFAULT_EMAIL:-pgadmin4@pgadmin.org}
+      PGADMIN_DEFAULT_PASSWORD: ${PGADMIN_DEFAULT_PASSWORD:-admin}
+    volumes:
+      - pgadmin:/root/.pgadmin
+    ports:
+      - "${PGADMIN_PORT:-5050}:80"
+    networks:
+      - postgres  
+    restart: unless-stopped
 
 networks:
   postgres:
     driver: bridge
- 
+
 volumes:
-    postgres:
-    pgadmin:
-    data:
+  postgres:
+  pgadmin:
+  mongo_data:
+  data:
+  db_data:
